# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWAPI Laravel API",
        "description": "API REST que integra la Star Wars API (SWAPI) con caché, reintentos y DTOs. Gestiona planetas y habitantes localmente con relaciones Eloquent.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API v1 – local"
        }
    ],
    "paths": {
...
        "/peliculas": {
            "get": {
                "tags": [
                    "Películas Locales"
                ],
                "summary": "Listar películas locales",
                "operationId": "a86b7f4eee1bd33dd8586df9648fce03",
                "responses": {
                    "200": {
                        "description": "Lista paginada de películas"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Películas Locales"
                ],
                "summary": "Crear una película local con sus relaciones",
                "operationId": "8d3d188ef5dd00a58e600a568b32dfc4",
                "responses": {
                    "201": {
                        "description": "Película creada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/peliculas/{id}": {
            "get": {
                "tags": [
                    "Películas Locales"
                ],
                "summary": "Obtener una película local con todas sus relaciones",
                "operationId": "26d9d19c84d30ae3661dd757e36a8406",
                "responses": {
                    "200": {
                        "description": "Película encontrada"
                    },
                    "404": {
                        "description": "Película no encontrada"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Películas Locales"
                ],
                "summary": "Actualizar una película local",
                "operationId": "d02ab2b8f2f3c77224ec0b739f465df7",
                "responses": {
                    "200": {
                        "description": "Película actualizada"
                    },
                    "404": {
                        "description": "Película no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Películas Locales"
                ],
                "summary": "Eliminar una película local",
                "operationId": "4ff38ced432a9d02a0b719f6b14d25ab",
                "responses": {
                    "200": {
                        "description": "Película eliminada"
                    },
                    "404": {
                        "description": "Película no encontrada"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key requerida. Configura API_KEY en tu .env y envíala como header X-API-KEY.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}